* Fixup the provisioning scripts and add a section for a clean ubuntu 16.04 machine.

+++ deploy/provision/dev.sh
@@ -12,7 +12,8 @@ if ! sudo -u postgres -- psql -tAc "SELECT 1 from pg_database WHERE datname='kal
 fi
 
 # Load Data
-source /home/vagrant/.zshrc
+source $HOME/.bashrc
+source $HOME/.virtualenvs/kala/bin/activate
 cd /srv/$PROJECT-app/django_$PROJECT
 python3 manage.py migrate
 echo "from auth.models import User; user = User.objects.create(**{'email': 'teststaff@example.com', 'is_active': True, 'is_staff': True, 'is_superuser': True, 'last_name': 'Staff', 'username': 'teststaff'}); user.set_password('test'); user.save();" | python3 manage.py shell
@@ -30,7 +31,10 @@ echo "from organizations.models import Organization; from projects.models import
 echo "from organizations.models import Organization; from projects.models import Project; from auth.models import User, Permissions, Permission; Permissions.objects.create(user=User.objects.get(email='testuser@example.com'), permission=Permission.objects.get(codename='delete_project'), object_uuid=Project.objects.get(name='Test Project').uuid)" | python3 manage.py shell
 
 # add the admin static pages.
-ln -s /home/vagrant/.virtualenvs/$PROJECT/lib/python3.5/site-packages/django/contrib/admin/static/admin /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static
-ln -s /home/vagrant/.virtualenvs/$PROJECT/lib/python3.5/site-packages/debug_toolbar/static/debug_toolbar /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static
-
+if ! [ -f /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static ]; then
+    ln -s /home/vagrant/.virtualenvs/$PROJECT/lib/python3.5/site-packages/django/contrib/admin/static/admin /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static
+fi
+if ! [ -f /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static ]; then
+    ln -s /home/vagrant/.virtualenvs/$PROJECT/lib/python3.5/site-packages/debug_toolbar/static/debug_toolbar /srv/$PROJECT-app/django_$PROJECT/django_$PROJECT/static
+fi
 service uwsgi restart
